Need to scrap this idea..
Want to keep a simple “ sign up to be contacted” form. Ask for type of
services requested (windows, gutters, other [please name]), name,
Email, phone, address, time to be contacted. Move the ‘Other Services”
Logos to about page. And the “referals” and the “sign up to receive
reminders” should be moved to home page. Perhaps as drop down menu
options?